feat(security): proteger endpoints con RBAC e integrar SecurityUtils
- Se implementó @PreAuthorize en los controladores para restringir el acceso basado en roles (RBAC).
- Se creó SecurityUtils para extraer la identidad y el gymId del usuario directamente desde el token JWT.
- Se refactorizo la logica de negocio de los use cases validando usuarios y permisos.
- Se dejaron TODOs por realizar en el siguiente commit (logica de negocio y validaciones con el TOKEN de JWT)

+++ src/main/java/com/trainingapp/trainingapp/application/usecase/exercise/DeleteExerciseUseCase.java
@@ -1,25 +1,55 @@
 package com.trainingapp.trainingapp.application.usecase.exercise;
 
 import com.trainingapp.trainingapp.domain.entity.exercise.Exercise;
+import com.trainingapp.trainingapp.domain.entity.user.Admin;
+import com.trainingapp.trainingapp.domain.entity.user.User;
+import com.trainingapp.trainingapp.domain.enums.user.Role;
 import com.trainingapp.trainingapp.domain.exception.exercise.ExerciseNotFoundException;
 import com.trainingapp.trainingapp.domain.repository.exercise.ExerciseRepository;
+import com.trainingapp.trainingapp.infrastructure.repository.jpa.config.security.SecurityUtils;
 import jakarta.transaction.Transactional;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 @Service
 public class DeleteExerciseUseCase {
 
     private final ExerciseRepository exerciseRepository;
+    private final SecurityUtils securityUtils;
 
-    public DeleteExerciseUseCase(ExerciseRepository exerciseRepository) {
+    public DeleteExerciseUseCase(ExerciseRepository exerciseRepository, SecurityUtils securityUtils) {
         this.exerciseRepository = exerciseRepository;
+        this.securityUtils = securityUtils;
     }
 
     @Transactional
     public void execute(Long id) {
-        Exercise exercise = exerciseRepository.findById(id)
-                .orElseThrow(() -> new ExerciseNotFoundException("The exercise with id " + id + " was not found."));
+        User currentUser = securityUtils.getCurrentUser();
+
+        Exercise exercise = findExerciseOrThrow(id);
+
+        validateDeletePermission(currentUser, exercise);
 
         exerciseRepository.delete(exercise);
+    }
+
+    private Exercise findExerciseOrThrow(Long id) {
+        return exerciseRepository.findById(id)
+                .orElseThrow(() -> new ExerciseNotFoundException(
+                        "The exercise with id " + id + " was not found."));
+    }
+
+    private void validateDeletePermission(User user, Exercise exercise) {
+        if (user.getRole() == Role.SUPER_ADMIN) return;
+
+        if (exercise.getIsBase()) {
+            throw new AccessDeniedException("No se pueden eliminar ejercicios base del sistema.");
+        }
+
+        // Sabemos que es ADMIN, ya que el controlador no permite otro rol.
+        Admin admin = (Admin) user;
+        if (!exercise.getGymId().equals(admin.getGymId())) {
+            throw new AccessDeniedException("No tienes permiso para eliminar ejercicios de otro gimnasio.");
+        }
     }
 }
